+++ src/Discord.Net.WebSocket/Entities/Messages/SocketReactionMessage.cs
@@ -1,0 +1,55 @@
+using Discord.Rest;
+using System.Collections.Generic;
+using System.Diagnostics;
+using System.Linq;
+using System.Threading.Tasks;
+
+namespace Discord.WebSocket
+{
+    /// <summary>
+    ///     Represents a WebSocket-based message that reactions can be added to.
+    /// </summary>
+    [DebuggerDisplay(@"{DebuggerDisplay,nq}")]
+    public abstract class SocketReactionMessage : SocketMessage, IReactionMessage
+    {
+        private readonly List<SocketReaction> _reactions = new List<SocketReaction>();
+
+        internal SocketReactionMessage(DiscordSocketClient discord, ulong id, ISocketMessageChannel channel, SocketUser author, MessageSource source)
+            : base(discord, id, channel, author, source)
+        {
+        }
+
+        /// <inheritdoc />
+        public IReadOnlyDictionary<IEmote, ReactionMetadata> Reactions => _reactions.GroupBy(r => r.Emote).ToDictionary(x => x.Key, x => new ReactionMetadata { ReactionCount = x.Count(), IsMe = x.Any(y => y.UserId == Discord.CurrentUser.Id) });
+
+        internal void AddReaction(SocketReaction reaction)
+        {
+            _reactions.Add(reaction);
+        }
+        internal void RemoveReaction(SocketReaction reaction)
+        {
+            if (_reactions.Contains(reaction))
+                _reactions.Remove(reaction);
+        }
+        internal void ClearReactions()
+        {
+            _reactions.Clear();
+        }
+
+        /// <inheritdoc />
+        public Task AddReactionAsync(IEmote emote, RequestOptions options = null)
+            => MessageHelper.AddReactionAsync(this, emote, Discord, options);
+        /// <inheritdoc />
+        public Task RemoveReactionAsync(IEmote emote, IUser user, RequestOptions options = null)
+            => MessageHelper.RemoveReactionAsync(this, user.Id, emote, Discord, options);
+        /// <inheritdoc />
+        public Task RemoveReactionAsync(IEmote emote, ulong userId, RequestOptions options = null)
+            => MessageHelper.RemoveReactionAsync(this, userId, emote, Discord, options);
+        /// <inheritdoc />
+        public Task RemoveAllReactionsAsync(RequestOptions options = null)
+            => MessageHelper.RemoveAllReactionsAsync(this, Discord, options);
+        /// <inheritdoc />
+        public IAsyncEnumerable<IReadOnlyCollection<IUser>> GetReactionUsersAsync(IEmote emote, int limit, RequestOptions options = null)
+            => MessageHelper.GetReactionUsersAsync(this, emote, limit, Discord, options);
+    }
+}
